requirements.txt
-e api
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
tqdm==4.43.0
Werkzeug==1.0.0
